org-remember: Skip creation of the default '* Notes' task
Remember and refile processing does not require a task.  This change
removes the unneeded default task.

This supports a workflow where new remember tasks and notes go into a
mostly empty file which just has #+FILETAGS: at the top and nothing
else.

This workflow has a minimal number of remember templates
  - one for new tasks and (filed in tasks.org)
  - one for new notes (filed in notes.org)
  - one for phone calls (filed in phone.org)

New tasks are added as top-level tasks to the end of these files and
the #+FILETAGS: REFILE header causes each task to be easy to find.
All tasks in these files are refiled to a more appropriate org file at
a later time.