Revert "Jump to selected clock task when starting the clock"
This reverts commit 61dfa0c67fcb47351f2139160ad7f59d7e742254.

This wasn't such a great idea after all.  There is one case where
this is really annoying - if you are in the middle of displaying
an agenda view with lots of filters active and you change what you
are clocking it jumps away and you lose the agenda view.

This one case makes hitting a dedicated function key that runs
org-clock-goto more desireable.