Make org table result have header row and hline when appropriate. (This should have been included with commit a39d68640e3fbc059f3b05c9eafc8bf07cc4b51d). The approach I've taken is to make R always write the column names into the tsv file. If the result does not have user-supplied column-names, then R writes default column names of the form V1,V2,...,V$numcols. After reading the tsv back into lisp, we check whether it has "real" column names or the R defaults. If it has real column names, then I give the org table an hline.
I expect this will conflict with Eric's work on sessions, but we can deal with that.

There is a remaining quoting issue that needs to be dealt
with. E.g. with an unquoted string in the input, it comes out as
quoted in the result. This seems bad, since if you use the output as
new input you get something different (with extra
quoting). (non-idempotence if that's the right term).

| col1 | col2    | col3 |
|------+---------+------|
|    1 | 2       |    3 |
|    4 | schulte |    6 |

tabel

| "col1" | "col2"    | "col3" |
|--------+-----------+--------|
|      1 | 2         |      3 |
|      4 | "schulte" |      6 |