Agenda: Allow custom command options to work better
When the agenda buffer for an agenda series is created, this was so
far done without the options of the custom agenda command in  place.
This meant that some options would not take effect, because the only
place where that did would have an effect was during creation of that
buffer.

This commit makes sure that the global options of an agenda series are
in fact in place when the buffer is created.