Initialize org-clock-notification-was-shown on clock-in
This makes org-clock-sound a little more well-behaved.  The clock
sound was only being issued if you were already clocking a task that
was under its limit, and changed to clocking something else.  Now each
time you clock in a new task it checks the estimated effort limit and
issues the org-clock-sound if the limit is exceeded.

This patch issues the org-clock-sound whenever you clock in a task
that is already over its estimated effort amount.  This allows you to
use the estimated effort as a limit for the amount of time to spend on
a task.  Each time you clock in the task you'll be notified that you
are over the limit.  This is great for tasks where you want to limit
the amount of time spent (per day, since last repeat) and is a
reminder to go work on something else instead.