LaTeX export: Allow postscript extensions for images
Some propel process LaTeX filed not directly to pdf, but go
through dvi and then to ps or pdf.  In that case, allowed images
are ps and eps files, not pdf and jpg.

This commit adds the two extensions, so that export using that
alternative path can be supported better.  However, it is up to the
user to make sure that the images are actually compatible with the
backend.