Revert "making :results replace the default"
This reverts commit a13cbf64b6462cf007632c68bf6cf8fc298e919a.

I'm in favour of this change, but it seems that we need a more sophisticated way of combining plists before we can
change the default header args in this way. Otherwise, we get two entries for :results (one from the defaults and one
from the header args), whereas what we want is a single entry for :results with space-separated values.