Archive subtree and move to next visible task
This skips over blank lines preceeding the next task when archiving a task or
subtree.  This allows us to use a keyboard macro to remember the archive key
sequence incantation and skips to the next heading so we don't accidentally
archive the wrong task.

This fixes an issue with blank line handling.  Tasks which have blank lines
following them do not include the blank lines -- they are part of the following
task.  This moves the point forward to the next visible headline so that if we
repeat the archive operation it affects the following task and not the previous
one by mistake.