Revert "Altering org-babel-execute:emacs-lisp in accordance with preceding changes."
This reverts commit d8001facab6ffa844ed5f042facd26268c9aec8c.

Going back to original plan of simply passing (cmd body params) to the org-babel-execute:LANG functions.

The benefit of this is that languages will have access to the full params list. A downside is that code parsing the
params list and referencing variables is currently duplicated across the languages, so perhaps we can aim to reduce
that code duplication at some point.