Adding support for column names (header line) when using R.
Unlike the other languages, it's central to R to be able to index
columns of a data frame d, either by d[,"columnname"] of d$columnname.

With this change, if colnames are present in the *input* from
org-babel, the corresponding R variable is *always* constructed with
the colnames.

In addition, with the :colnames header arg, the *output* to elisp/org
buffer contains the colnames separated from the rest of the table by
'hline. This behaviour is not default because other languages may
expect a simple table without the 'hline.