Bugfix about using the calendar on a separate frame.
`org-read-date' was loosing the focus when the calendar was displayed on
a separate frame.  This patch by Robert Goldman solves it by introducing
a new macro `org-save-frame-excursion' which preserve the frame focus.

See this thread: http://article.gmane.org/gmane.emacs.orgmode/15528

(This fixes the issue I was trying to fix in a previous commit.)