Speed-up cycling in buffers with *many* siblings, and *many* drawers
For this we took another look at when drawers actually have to be
hidden again and found that CONTENTS view does not need it, and that
CHILDREN view only needs it before the first child.

The second speed-up comes from advising outline-end-of-subtree to use
the Org version when in org-mode.

The third speed-up comes from using a better way to find the next
visible line, using `next-single-char-property-change'.

Finally, `org-forward-same-level' and `org-backward-same-level' are
faster versions of their outline equivalents and are now bound to
`C-c C-f' and `C-c C-b'.