LaTeX export: Remove bug in table export.
A `save-excursion' around a call to org-table-align make point end up
*before* the table.  The reason is that a table align replaces the
entire table, including the newline before it.  When the table is
removed in order to be replaced, the marker created by
`save-excursion' slips.  `org-table-align' has it's own, built-in
`save-excursion' by remembering the line and column where the cursor
was before the align.