No longer loading all languages by default.  REQUIRES A CHANGE TO YOUR CONFIG!
This fixes a bug: DONE require users to explicitly turn on each language

as we continue adding more languages (each with it's own set of major
mode requirements), we are going to want users to only turn on the
languages that they intend to use.

See the install instructions in org-babel-worg.org, also take a look
at the "Requirements" sections of org-babel-ruby.el and
org-babel-gnuplot.el for pointers to downloading and installing their
requirements (which are no longer distributed in the util directory).