Export: Remove table rows that only contain width and alignment markers
The width and alignment in table columns can be set with a cookie like
"<10>" or "<r>" or "<r10>".  In order to keep Org from exporting such
lines, the first column of a line can contain only "/".  However, for
convenience, this commit implements a special case:  If the entire row
contains only sch markers, the line will automatically be discarded
during export.