Optionally prevent altering indentation of source code.
New customizable variable org-src-preserve-indentation, if non-nil,
prevents automatic removal of rectangular blocks of leading whitespace
in source code blocks. This is necessary when embedding python code in
an org file with a view to extracting all python blocks into a single
file (because whitespace indentation is critical in python code). The
changes in this commit also fix two bugs:

- Setting org-edit-src-content-indentation to zero resulted in an
  infinite loop in org-edit-src-exit.

- Fixed width regions are defined by lines starting with a colon, with
  optional leading whitespace. However, if there was any leading
  whitespace, org-edit-src-exit behaved incorrectly, causing the leading
  whitespace to grow.