Provide frame/window control when switching to source code edit buffer.
These changes provides frame / window management preferences for the
edit buffer generated using C-c ' on a source code block. Preferences
are specified by a new variable org-src-window-setup, which is based
on org-agenda-window-setup and has the same four options:
current-window, other-window, reorganize-frame, other-frame. These
behave as follows:

* current-window
The edit buffer appears in the current window.

* other-window
`switch-to-buffer-other-window' is used to switch to the edit buffer
in the same frame.

* reorganize-frame [default]
The current frame is reorganized so that it is split between the
source code edit buffer and the parent org buffer.

* other-frame
The edit buffer appears in a new frame. That frame is deleted when
exiting the edit buffer with C-c '.