New faces for title, date, author and email address lines.
By default, title, author, date and email lines appear in dark blue
with the initial keywords greyed out. The title is in a larger font
than the others. This is implemented by the following new faces:

org-document-title
org-document-info
org-document-info-keyword

In addition, the variable org-hidden-keywords can be used to make the
corresponding keywords disappear.