`org-timer-set-timer': new prefix argument.
Called with a numeric prefix argument, `org-timer-set-timer' uses
this numeric value as the duration of the timer.

Called with a `C-u' prefix argument, use `org-timer-default-timer'
without prompting the user for a duration.

With two `C-u' prefix arguments, use `org-timer-default-timer'
without prompting the user for a duration and automatically
replace any running timer."