Reorder Makefile
OK, we finally have a make file where I can do a complete release,
including all the git updates on master and maint branches, with a
single command:

  make release TAG=N.NN

It is important that the working directory is clean when calling this
command.

Since this command can fail and then leave the git repo in an
uncertain state, it is best to first run this:

  make testrelease

This will create a throw-away branch "testrelease" and try alll the
commands there, to make sure everything is fine.