Provide link property for message-id without angle brackets
* org-wl.el (org-wl-store-link-message): Provide link property
for message-id without angle brackets.

The message-id header field can be used to maintain the connection
between an Org entry and an internet message.  The angle brackets have
a special meaning when performing a TAGS/PROPS/TODO query (Cf. Org
mode manual 10.3.3).  To be able to use the message-id as an entry
property we have to remove the angle brackets to be able to perform a
search for the entry associated with a particular message.